Move postgres tests from test_operators to test_postgres
